requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.3
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.3
orjson==3.10.12
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pillow==10.4.0
pinecone==5.4.1
pinecone-client==2.2.4
pinecone-plugin-inference==3.0.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.1.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.28.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
streamlit==1.29.0
sympy==1.13.1
tenacity==8.5.0
tesseract==0.1.3
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
